+++ express-zod-api/src/endpoint.ts
@@ -1,7 +1,6 @@
 import { Request, Response } from "express";
 import * as R from "ramda";
-import { z, globalRegistry } from "zod/v4";
-import type { $ZodObject } from "zod/v4/core";
+import { z, globalRegistry } from "zod";
 import { NormalizedResponse, ResponseVariant } from "./api-response";
 import { findRequestTypeDefiningSchema } from "./deep-checks";
 import {
@@ -22,7 +21,7 @@ import { IOSchema } from "./io-schema";
 import { lastResortHandler } from "./last-resort";
 import { ActualLogger } from "./logger-helpers";
 import { LogicalContainer } from "./logical-container";
-import { getBrand, getExamples } from "./metadata";
+import { getBrand } from "./metadata";
 import { ClientMethod, CORSMethod, Method, SomeMethod } from "./method";
 import { AbstractMiddleware, ExpressMiddleware } from "./middleware";
 import { ContentType } from "./content-type";
@@ -86,14 +85,14 @@ export class Endpoint<
 
   /** considered expensive operation, only required for generators */
   #ensureOutputExamples = R.once(() => {
-    if (getExamples(this.#def.outputSchema).length) return; // has examples on the output schema, or pull up:
-    if (!isSchema<$ZodObject>(this.#def.outputSchema, "object")) return;
-    const examples = pullResponseExamples(this.#def.outputSchema as $ZodObject);
+    if (globalRegistry.get(this.#def.outputSchema)?.examples?.length) return; // examples on output schema, or pull up:
+    if (!isSchema<z.core.$ZodObject>(this.#def.outputSchema, "object")) return;
+    const examples = pullResponseExamples(this.#def.outputSchema);
     if (!examples.length) return;
     const current = this.#def.outputSchema.meta();
     globalRegistry
       .remove(this.#def.outputSchema) // reassign to avoid cloning
-      .add(this.#def.outputSchema as $ZodObject, { ...current, examples });
+      .add(this.#def.outputSchema, { ...current, examples });
   });
 
   constructor(def: {